+++ asmifier-plugin/src/main/java/com/likethesalad/asm/tasks/AsmifierTask.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
+import javax.inject.Inject;
 import org.gradle.api.DefaultTask;
 import org.gradle.api.Project;
 import org.gradle.api.file.ConfigurableFileCollection;
@@ -16,6 +17,7 @@ import org.gradle.api.tasks.PathSensitive;
 import org.gradle.api.tasks.PathSensitivity;
 import org.gradle.api.tasks.SkipWhenEmpty;
 import org.gradle.api.tasks.TaskAction;
+import org.gradle.process.ExecOperations;
 import org.gradle.work.ChangeType;
 import org.gradle.work.FileChange;
 import org.gradle.work.InputChanges;
@@ -33,6 +35,10 @@ public abstract class AsmifierTask extends DefaultTask {
   @OutputDirectory
   public abstract DirectoryProperty getOutputDir();
 
+  @SuppressWarnings("JavaxInjectOnAbstractMethod")
+  @Inject
+  public abstract ExecOperations getExecOperations();
+
   @TaskAction
   public void execute(InputChanges inputChanges) {
     Project project = getProject();
@@ -46,16 +52,15 @@ public abstract class AsmifierTask extends DefaultTask {
         }
 
         if (fileChange.getChangeType().equals(ChangeType.REMOVED)) {
-          File targetFile =
-              getTargetFile(fileChange.getNormalizedPath(), fileChange.getFile().getName());
+          File targetFile = getTargetFile(fileChange.getNormalizedPath());
           if (!targetFile.delete()) {
             getLogger().warn("Could not delete: {}", targetFile);
           }
         } else if (fileChange.getChangeType().equals(ChangeType.ADDED)
             || fileChange.getChangeType().equals(ChangeType.MODIFIED)) {
           String relativeSourcePath = fileChange.getNormalizedPath();
-          File outputFile = getOutputFile(relativeSourcePath, fileChange.getFile().getName());
-          asmifyToFile(outputFile, project, classpath, relativeSourcePath);
+          File outputFile = getOutputFile(relativeSourcePath);
+          asmifyToFile(outputFile, classpath, relativeSourcePath);
         }
       }
     } else {
@@ -65,52 +70,52 @@ public abstract class AsmifierTask extends DefaultTask {
               fileVisitDetails -> {
                 if (!fileVisitDetails.isDirectory()) {
                   String relativeSourcePath = fileVisitDetails.getRelativePath().getPathString();
-                  File outputFile = getOutputFile(relativeSourcePath, fileVisitDetails.getName());
-                  asmifyToFile(outputFile, project, classpath, relativeSourcePath);
+                  File outputFile = getOutputFile(relativeSourcePath);
+                  asmifyToFile(outputFile, classpath, relativeSourcePath);
                 }
               });
     }
   }
 
-  private void asmifyToFile(
-      File outputFile, Project project, FileCollection classpath, String relativeSourcePath) {
+  private void asmifyToFile(File outputFile, FileCollection classpath, String relativeSourcePath) {
     try (FileOutputStream outputStream = new FileOutputStream(outputFile)) {
-      asmify(project, classpath, relativeSourcePath, outputStream);
+      asmify(classpath, relativeSourcePath, outputStream);
     } catch (IOException e) {
-      throw new RuntimeException(e);
+      throw new RuntimeException(
+          "Exception during asmifier run where the source is: "
+              + relativeSourcePath
+              + " and the output is: "
+              + outputFile.getPath(),
+          e);
     }
   }
 
-  private File getOutputFile(String sourceRelativePath, String sourceFileName) {
-    File outputFile = getTargetFile(sourceRelativePath, sourceFileName);
+  private File getOutputFile(String sourceRelativePath) {
+    File outputFile = getTargetFile(sourceRelativePath);
     try {
       outputFile.getParentFile().mkdirs();
       outputFile.createNewFile();
     } catch (IOException e) {
-      throw new RuntimeException(e);
+      throw new RuntimeException("Exception while getting output file: " + outputFile.getPath(), e);
     }
     return outputFile;
   }
 
-  private File getTargetFile(String sourceRelativePath, String sourceFileName) {
-    String classSimpleName = sourceFileName.replaceFirst("\\.class", "");
-    return getOutputDir()
-        .file(
-            "asm/" + sourceRelativePath.replaceFirst(sourceFileName, classSimpleName + "Dump.java"))
-        .get()
-        .getAsFile();
+  private File getTargetFile(String sourceRelativePath) {
+    String newPath = sourceRelativePath.replaceFirst("\\.class", "Dump.java");
+    return getOutputDir().file("asm/" + newPath).get().getAsFile();
   }
 
-  private void asmify(
-      Project project, FileCollection classpath, String relativePath, OutputStream outputStream) {
+  private void asmify(FileCollection classpath, String relativePath, OutputStream outputStream) {
     String className = relativePath.replaceFirst("\\.class", "").replaceAll("/", ".");
-    project.javaexec(
-        javaExecSpec -> {
-          javaExecSpec.setClasspath(classpath);
-          javaExecSpec.getMainClass().set("org.objectweb.asm.util.ASMifier");
-          javaExecSpec.args("-nodebug");
-          javaExecSpec.args(className);
-          javaExecSpec.setStandardOutput(outputStream);
-        });
+    getExecOperations()
+        .javaexec(
+            javaExecSpec -> {
+              javaExecSpec.setClasspath(classpath);
+              javaExecSpec.getMainClass().set("org.objectweb.asm.util.ASMifier");
+              javaExecSpec.args("-nodebug");
+              javaExecSpec.args(className);
+              javaExecSpec.setStandardOutput(outputStream);
+            });
   }
 }