+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Gradle ASMifier Changelog
 
+## Unreleased
+
+- Adding support for anonymous classes.
+- Removing usage of deprecated tools.
+- Ensuring only .class files are targeted.
+
 ## Version 1.1.0 (2025-02-27)
 
 - Making inputs flexible